requirements.txt
fastapi
uvicorn
pydantic
firebase-admin
google-api-python-client
google-auth
